requirements.txt
torch
torchvision
numpy
tqdm
Pillow
scikit-learn
scipy
matplotlib
